Add a JSP pre-compilation process to maximize performance
Some application servers have this capability of pre-compiling JSPs ahead of time, however WildFly/JBoss only compile .jsp files on-the-fly, in the first servlet request.

To address that, the JSP compilation tasks are asynchronously scheduled during the application deployment step by making one lightweight HTTP request for each route (JAX-RS Resource / Servlet) that invoke the JSP lifecycle.

The only "pitfall" of this solution is: "JSPC requests" that are meant to only invoke the JSP compiler must not perform business logic or anything that can slow down the compiler invocation, following this simple rule we won't have any issues, btw that logic was abstracted in the View objects =)